requirements.txt
discord.py
datetime
aiohttp
hyper-status
youtube-dl
